requirements.txt
ansible-core==2.16.2
wheel
openstacksdk
flake8
pytest
